+++ src/validators/object/object.validator.js
@@ -1,5 +1,6 @@
 const _ = require('lodash')
 const { getCurrentTranslation } = require('../../translation')
+const { createContext } = require('../../context')
 
 const objectValidator = {
   $type: 'object',
@@ -17,18 +18,46 @@ const objectValidator = {
     const { getValidators } = require('../validators')
 
     const schemaObjProps = Object.keys(ctx.schema.props)
+    const valueProps = Object.keys(ctx.value)
+
+    if (ctx.schema.strict) {
+      const extraProps = valueProps.filter(prop => schemaObjProps.indexOf(prop) < 0)
+
+      if (extraProps.length > 0) {
+        ctx.registerErrors({
+          strict: {
+            message: getCurrentTranslation().validators.object.errors.strict(extraProps),
+            value: ctx.value,
+            extraProps
+          }
+        })
+        return 
+      }
+    }
 
     for (let propKey of schemaObjProps) {
       const propSchema = ctx.schema.props[propKey]
       const propValue = ctx.value[propKey]
       const validator = getValidators()[propSchema.$type]
 
-      ctx.addToPath(propKey)
-      validator.validate({
-        ...ctx,
+      const propContext = createContext({
         schema: propSchema,
         value: propValue
       })
+
+      ctx.addToPath(propKey)
+      validator.validate(propContext)
+      
+      propContext.path.forEach((path, index) => {
+        if (index > 0) { // skip $root
+          ctx.addToPath(path)
+        }
+      })
+
+      if (propContext.errors.$root) {
+        ctx.registerErrors(propContext.errors.$root)
+      }
+
       ctx.popPath()
     }
   }