+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Adding `password` validator
+- Adding `url` validator
 
 ## [1.6.0] - 2019-02-04