+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added - 2019-02-16
+## [1.8.0] - 2019-02-16
+
+### Added
 
 - Adding `strict` mode to `object` validator.
 - Adding `throw` middleware.